make delete conferenceroom command

+++ ConferenceRoomBooking.Application/Features/ConferenceRooms/Handlers/Commands/DeleteConferenceRoomCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using ConferenceRoomBooking.Application.Contracts.Repositories;
+using ConferenceRoomBooking.Application.DTOs.ConferenceRoomRequest;
+using ConferenceRoomBooking.Application.Exceptions;
 using ConferenceRoomBooking.Application.Features.ConferenceRooms.Requests.Commands;
 using ConferenceRoomBooking.Application.Responces;
 using ConferenceRoomBooking.Domain.Entities;
@@ -19,7 +21,18 @@ namespace ConferenceRoomBooking.Application.Features.ConferenceRooms.Handlers.Co
 
         public async Task<Result> Handle(DeleteConferenceRoomCommand request, CancellationToken cancellationToken)
         {
-            var conferenceRoom = new ConferenceRoom() { Id = request.Id };
+            var conferenceRoomFilter = new ConferenceRoomFilterDto() { Guids = [request.Id] };
+            var ConferenceRoomResult = await _conferenceRoomRepository.GetAsync(conferenceRoomFilter);
+
+            var conferenceRoom = ConferenceRoomResult.MatchSuccess(
+                conferenceRooms => conferenceRooms.FirstOrDefault()
+            );
+
+            if (conferenceRoom == null)
+            {
+                return new Result<ConferenceRoomDto>(new NotFoundException(nameof(ConferenceRoom)));
+            }
+
             var deletedConferenceRoomResult = await _conferenceRoomRepository.DeleteAsync(conferenceRoom);
 
             return deletedConferenceRoomResult;
